feat: add `restart` option to Win+P display menu
this restarts the display manager. i havent exactly
figured out why this is needed but its usually required
when using the hdmi switcher. x sorta-but-not-really
detects monitors changing, this forces it to actually
detect when they change.

there is probably a better way to do this without restarting the
entire display manager...